refactored: renamed chat_info to info
