+++ src/model/chat/info.rs
